rooms route added in spec Dashboard
