+++ src/styles/Dashboard.module.scss
@@ -1798,6 +1798,7 @@ body.activemodal {
 }
 
 .temperarly_added{
+  width: 100%;
   height: 200px;
   display: flex;
   justify-content: center;
